utils: move top-level, keep old module available for now
