+++ src/abstract/modular.ts
@@ -14,7 +14,7 @@ import {
   ensureBytes,
   numberToBytesBE,
   numberToBytesLE,
-} from './utils.ts';
+} from '../utils.ts';
 
 // prettier-ignore
 const _0n = BigInt(0), _1n = BigInt(1), _2n = /* @__PURE__ */ BigInt(2), _3n = /* @__PURE__ */ BigInt(3);
